* Moved all interfaces into the Contracts library * Renamed Month enum to Months to help distinguish it from a model file
