+++ ScheduledWorker.Library.Contracts/Months.cs
@@ -1,0 +1,61 @@
+﻿namespace ScheduledWorker.Library.Contracts
+{
+    /// <summary>
+    /// This enumerates all the months in a year.
+    /// </summary>
+    public enum Months
+    {
+        /// <summary>
+        /// The month is unknown or unspecified.
+        /// </summary>
+        None = 0,
+        /// <summary>
+        /// The first month of the year.
+        /// </summary>
+        January = 1,
+        /// <summary>
+        /// The second month of the year.
+        /// </summary>
+        February = 2,
+        /// <summary>
+        /// The third month of the year.
+        /// </summary>
+        March = 3,
+        /// <summary>
+        /// The fourth month of the year.
+        /// </summary>
+        April = 4,
+        /// <summary>
+        /// The fifth month of the year.
+        /// </summary>
+        May = 5,
+        /// <summary>
+        /// The sixth month of the year.
+        /// </summary>
+        June = 6,
+        /// <summary>
+        /// The seventh month of the year.
+        /// </summary>
+        July = 7,
+        /// <summary>
+        /// The eighth month of the year.
+        /// </summary>
+        August = 8,
+        /// <summary>
+        /// The ninth month of the year.
+        /// </summary>
+        September = 9,
+        /// <summary>
+        /// The tenth month of the year.
+        /// </summary>
+        October = 10,
+        /// <summary>
+        /// The eleventh month of the year.
+        /// </summary>
+        November = 11,
+        /// <summary>
+        /// The twelfth month of the year.
+        /// </summary>
+        December = 12
+    }
+}
