+++ ScheduledWorker.Library.Contracts/Schedule/ISchedule.cs
@@ -1,0 +1,15 @@
+﻿namespace ScheduledWorker.Library.Contracts.Schedule
+{
+    using System.Collections.Generic;
+
+    /// <summary>
+    /// This outlines the methods and properties for a schedule.
+    /// </summary>
+    public interface ISchedule
+    {
+        /// <summary>
+        /// Gets the items scheduled to run.
+        /// </summary>
+        ICollection<IScheduleItem> Items { get; }
+    }
+}
