+++ ScheduledWorker.Library.Contracts/ScheduledWorker.Library.Contracts.csproj
@@ -41,9 +41,21 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Intervals\IInterval.cs" />
+    <Compile Include="IScheduleManager.cs" />
     <Compile Include="Logging\ILogger.cs" />
     <Compile Include="Logging\LoggingLevels.cs" />
+    <Compile Include="Months.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Schedule\IDailyScheduleItem.cs" />
+    <Compile Include="Schedule\IMonthlyScheduleItem.cs" />
+    <Compile Include="Schedule\IRunNowScheduleItem.cs" />
+    <Compile Include="Schedule\ISchedule.cs" />
+    <Compile Include="Schedule\IScheduleItem.cs" />
+    <Compile Include="Schedule\IScheduleSettings.cs" />
+    <Compile Include="Schedule\IWeeklyScheduleItem.cs" />
+    <Compile Include="WorkerTaskEventArgs.cs" />
+    <Compile Include="Worker\IWorkerTask.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
