+++ ScheduledWorker.Library/ScheduledWorker.Library.csproj
@@ -53,21 +53,9 @@
     <Compile Include="Configuration\Weekly\WeeklyScheduleCollection.cs" />
     <Compile Include="Configuration\Weekly\WeeklyScheduleItem.cs" />
     <Compile Include="Configuration\WorkerTaskConverter.cs" />
-    <Compile Include="Contracts\Schedule\IDailyScheduleItem.cs" />
-    <Compile Include="Contracts\Intervals\IInterval.cs" />
-    <Compile Include="Contracts\Schedule\IMonthlyScheduleItem.cs" />
-    <Compile Include="Contracts\Schedule\IRunNowScheduleItem.cs" />
-    <Compile Include="Contracts\Schedule\ISchedule.cs" />
-    <Compile Include="Contracts\Schedule\IScheduleSettings.cs" />
-    <Compile Include="Contracts\Schedule\IWeeklyScheduleItem.cs" />
-    <Compile Include="Contracts\WorkerTaskEventArgs.cs" />
     <Compile Include="Core\Intervals\BaseInterval.cs" />
     <Compile Include="Core\Intervals\MinuteInterval.cs" />
     <Compile Include="Core\Intervals\SecondInterval.cs" />
-    <Compile Include="IScheduleManager.cs" />
-    <Compile Include="Contracts\Month.cs" />
-    <Compile Include="Contracts\Schedule\IScheduleItem.cs" />
-    <Compile Include="Contracts\Worker\IWorkerTask.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ScheduleManager.cs" />
     <Compile Include="Worker\BaseWorkerTask.cs" />
